moved RasterExpression components
